Moved screen classes to a separate file.
